requirements.txt
APScheduler
Jinja2
PyMySQL
python-dateutil
requests
GitPython
SQLAlchemy
uvicorn
fastapi
werkzeug
python-multipart